refactor(ds): update components
