+++ ui/design-system-core/src/components/AppIcon/icon-by-type.tsx
@@ -1,16 +1,17 @@
 import React from 'react';
-import Icon from '../Icon';
-import { AppIconProps } from './index';
-import { LogoTypeSource } from '@akashaorg/typings/ui';
 import { apply, tw } from '@twind/core';
 
-type AppImgProps = {
-  appImg: AppIconProps['appImg'];
-  size: AppIconProps['size'];
-  placeholderIconType: AppIconProps['placeholderIconType'];
-  breakPointSize?: AppIconProps['breakPointSize'];
-  accentColor?: AppIconProps['accentColor'];
-  color?: AppIconProps['iconColor'];
+import { LogoTypeSource } from '@akashaorg/typings/ui';
+
+import Icon from '../Icon';
+import { AppIconProps } from '.';
+import { Color } from '../types/common.types';
+
+export type AppImgProps = Pick<
+  AppIconProps,
+  'appImg' | 'size' | 'placeholderIconType' | 'breakPointSize' | 'accentColor'
+> & {
+  color?: Color;
 };
 
 const IconByType: React.FC<AppImgProps> = ({
@@ -34,7 +35,13 @@ const IconByType: React.FC<AppImgProps> = ({
   }
 
   if (appImg?.type === (LogoTypeSource.String || LogoTypeSource.IPFS)) {
-    return <img className={tw(apply`rounded-[50%] object-contain`)} src={appImg?.value} />;
+    return (
+      <img
+        alt={appImg.type}
+        className={tw(apply`rounded-[50%] object-contain`)}
+        src={appImg?.value}
+      />
+    );
   }
   return (
     <Icon
